university-web-demo/requirements.txt
Flask==3.0.3
pyodbc==5.2.0
gunicorn==23.0.0
python-decouple==3.8